reorder teleop, move reset state
