+++ lib/routes/auto/AutonomousDataRoute.dart
@@ -25,6 +25,7 @@ import 'package:scouting_platform/utils/data/values/EndgameValues.dart';
 import 'package:scouting_platform/utils/data/values/PrematchValues.dart';
 import 'package:scouting_platform/utils/data/values/SettingValues.dart';
 import 'package:scouting_platform/utils/data/values/TeleoperatedValues.dart';
+import 'package:scouting_platform/utils/helpers/AppDataHelper.dart';
 import 'package:scouting_platform/utils/helpers/ScheduleHelper.dart';
 import 'package:scouting_platform/utils/helpers/UIHelper.dart';
 
@@ -88,13 +89,27 @@ class _DataRouteState extends State<AutonomousDataRoute> {
                     Row(
                       mainAxisAlignment: MainAxisAlignment.end,
                       children: [
-                        // TitleStyle(
-                        //     text: "L4",
-                        //     padding: const EdgeInsets.only(
-                        //         right: 0.0, bottom: 10.0)),
-                        SizedBox(
-                          width: 120.0,
-                        ),
+                        // reset button
+                        ElevatedButton(
+                            style: ElevatedButton.styleFrom(
+                              minimumSize: const Size(150.0, 37.0),
+                              backgroundColor: AppStyle.textInputColor,
+                            ),
+                            onPressed: () {
+                              Navigator.push(context,
+                                  MaterialPageRoute(builder: (context) {
+                                return const PrematchRoute(title: "Prematch");
+                              }));
+                            },
+                            child: const Text(
+                              "< Prematch",
+                              style: TextStyle(
+                                fontSize: 16.0,
+                                fontFamily: "Helvetica",
+                                color: Colors.white,
+                              ),
+                            )),
+                        SizedBox(width: 8.0),
 
                         // reset button
                         ElevatedButton(
@@ -115,7 +130,7 @@ class _DataRouteState extends State<AutonomousDataRoute> {
                           ),
                         ),
 
-                        const SizedBox(width: 16.0),
+                        const SizedBox(width: 8.0),
 
                         // next page button
                         ElevatedButton(
@@ -182,90 +197,7 @@ class _DataRouteState extends State<AutonomousDataRoute> {
         }
 
         setState(() {
-          // main
-          AutonomousValues.autoStartPosition.text = "center";
-          AutonomousValues.autoMobility.text = "no";
-          AutonomousValues.coralMissed.text = "0";
-          AutonomousValues.coralHP.text = "0";
-          AutonomousValues.coralField.text = "0";
-          AutonomousValues.algaeRemoved.text = "0";
-          AutonomousValues.algaeBarge.text = "0";
-          AutonomousValues.algaeProcessor.text = "0";
-
-          // tough coral
-          AutonomousValues.l1E.text = "0";
-          AutonomousValues.l1NE.text = "0";
-          AutonomousValues.l1NW.text = "0";
-          AutonomousValues.l1SE.text = "0";
-          AutonomousValues.l1SW.text = "0";
-          AutonomousValues.l1W.text = "0";
-
-          // L2 coral
-          AutonomousValues.l2A.text = "0";
-          AutonomousValues.l2B.text = "0";
-          AutonomousValues.l2C.text = "0";
-          AutonomousValues.l2D.text = "0";
-          AutonomousValues.l2E.text = "0";
-          AutonomousValues.l2F.text = "0";
-          AutonomousValues.l2G.text = "0";
-          AutonomousValues.l2H.text = "0";
-          AutonomousValues.l2I.text = "0";
-          AutonomousValues.l2J.text = "0";
-          AutonomousValues.l2K.text = "0";
-          AutonomousValues.l2L.text = "0";
-
-          // L3 coral
-          AutonomousValues.l3A.text = "0";
-          AutonomousValues.l3B.text = "0";
-          AutonomousValues.l3C.text = "0";
-          AutonomousValues.l3D.text = "0";
-          AutonomousValues.l3E.text = "0";
-          AutonomousValues.l3F.text = "0";
-          AutonomousValues.l3G.text = "0";
-          AutonomousValues.l3H.text = "0";
-          AutonomousValues.l3I.text = "0";
-          AutonomousValues.l3J.text = "0";
-          AutonomousValues.l3K.text = "0";
-          AutonomousValues.l3L.text = "0";
-
-          // L4 coral
-          AutonomousValues.l4A.text = "0";
-          AutonomousValues.l4B.text = "0";
-          AutonomousValues.l4C.text = "0";
-          AutonomousValues.l4D.text = "0";
-          AutonomousValues.l4E.text = "0";
-          AutonomousValues.l4F.text = "0";
-          AutonomousValues.l4G.text = "0";
-          AutonomousValues.l4H.text = "0";
-          AutonomousValues.l4I.text = "0";
-          AutonomousValues.l4J.text = "0";
-          AutonomousValues.l4K.text = "0";
-          AutonomousValues.l4L.text = "0";
-
-          TeleoperatedValues.coralNearL1.text = "0";
-          TeleoperatedValues.coralNearL2.text = "0";
-          TeleoperatedValues.coralNearL3.text = "0";
-          TeleoperatedValues.coralNearL4.text = "0";
-          TeleoperatedValues.coralFarL1.text = "0";
-          TeleoperatedValues.coralFarL2.text = "0";
-          TeleoperatedValues.coralFarL3.text = "0";
-          TeleoperatedValues.coralFarL4.text = "0";
-          TeleoperatedValues.coralMissed.text = "0";
-          TeleoperatedValues.algaeRemoved.text = "0";
-          TeleoperatedValues.algaeProcessor.text = "0";
-          TeleoperatedValues.algaeBarge.text = "0";
-          TeleoperatedValues.humanPlayerMisses.text = "0";
-          TeleoperatedValues.fieldCrosses.text = "0";
-          AutonomousValues.autoMobility.text = "No";
-          EndgameValues.endgame.text = "No"; // was climb, now endgame
-          EndgameValues.climbTime.text = "0"; // parked was below
-          EndgameValues.stopwatchState.text = "0";
-          EndgameValues.stopwatch.stop();
-          EndgameValues.stopwatch.reset();
-          CommentValues.autoComments.text = "";
-          CommentValues.autoOrder.text = "";
-          CommentValues.teleopComments.text = "";
-          CommentValues.endgameComments.text = "";
+          AppDataHelper.resetStates();
         });
         Navigator.of(context, rootNavigator: true).pop('dialog');
         Navigator.push(context, MaterialPageRoute(builder: (context) {
