+++ lib/routes/teleop/labels/TeleoperatedLabels2.dart
@@ -23,7 +23,7 @@ class _TeleoperatedLabels2State extends State<TeleoperatedLabels2> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "Coral L1 Far Scored",
+                "L2 Coral Near",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
@@ -39,7 +39,7 @@ class _TeleoperatedLabels2State extends State<TeleoperatedLabels2> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "Coral L2 Far Scored",
+                "L2 Coral Far",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
@@ -55,39 +55,7 @@ class _TeleoperatedLabels2State extends State<TeleoperatedLabels2> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "Coral L3 Far Scored",
-                textAlign: TextAlign.left,
-                style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 15.0),
-              ),
-            )),
-        // coral l4 far scored
-        Align(
-            alignment: Alignment.centerLeft,
-            child: Container(
-              width: 170.0,
-              padding:
-                  const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
-              child: const Text(
-                "Coral L4 Far Scored",
-                textAlign: TextAlign.left,
-                style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 15.0),
-              ),
-            )),
-        // algae barge
-        Align(
-            alignment: Alignment.centerLeft,
-            child: Container(
-              width: 170.0,
-              padding:
-                  const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
-              child: const Text(
-                "Algae Barge",
+                "Algae Removed",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
