+++ lib/routes/teleop/labels/TeleoperatedLabels4.dart
@@ -1,21 +1,36 @@
 // ignore_for_file: file_names
 import 'package:flutter/material.dart';
 
-class TeleoperatedLabels3 extends StatefulWidget {
-  const TeleoperatedLabels3({
+class TeleoperatedLabels4 extends StatefulWidget {
+  const TeleoperatedLabels4({
     super.key,
   });
 
   @override
-  State<TeleoperatedLabels3> createState() => _TeleoperatedLabels3State();
+  State<TeleoperatedLabels4> createState() => _TeleoperatedLabels4State();
 }
 
-class _TeleoperatedLabels3State extends State<TeleoperatedLabels3> {
+class _TeleoperatedLabels4State extends State<TeleoperatedLabels4> {
   @override
   Widget build(BuildContext context) {
     return Row(
       children: <Widget>[
-        // algae removed
+        //coral l3
+        Align(
+            alignment: Alignment.centerLeft,
+            child: Container(
+              width: 170.0,
+              padding: const EdgeInsets.only(left: 20.0, top: 20.0),
+              child: const Text(
+                "L4 Coral Near",
+                textAlign: TextAlign.left,
+                style: TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                    fontSize: 15.0),
+              ),
+            )),
+        // coral l4
         Align(
             alignment: Alignment.centerLeft,
             child: Container(
@@ -23,23 +38,7 @@ class _TeleoperatedLabels3State extends State<TeleoperatedLabels3> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "L3 Coral Near",
-                textAlign: TextAlign.left,
-                style: TextStyle(
-                    color: Colors.white,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 15.0),
-              ),
-            )),
-        // algae processor
-        Align(
-            alignment: Alignment.centerLeft,
-            child: Container(
-              width: 170.0,
-              padding:
-                  const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
-              child: const Text(
-                "L3 Coral Far",
+                "L4 Coral Far",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
@@ -54,7 +53,7 @@ class _TeleoperatedLabels3State extends State<TeleoperatedLabels3> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "Algae Processor",
+                "Algae Barge",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
@@ -69,7 +68,7 @@ class _TeleoperatedLabels3State extends State<TeleoperatedLabels3> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "HP Misses",
+                "Field Crosses",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
@@ -84,7 +83,7 @@ class _TeleoperatedLabels3State extends State<TeleoperatedLabels3> {
               padding:
                   const EdgeInsets.only(left: 20.0, top: 20.0, right: 30.0),
               child: const Text(
-                "Climb Time",
+                "Endgame",
                 textAlign: TextAlign.left,
                 style: TextStyle(
                     color: Colors.white,
